// apps/stock/src/types/stock.ts
export interface MotifData {
  skip?: boolean;
  id: string;
  dateValue: number;
  motifName: string;
  planeNum: string;
  droneType: string;
  fileName: string;
  height: string;
  width: string;
  depth: string;
  length: string;
  season: string;
  category: string;
  popular: string;
}

export interface TransitionData {
  id: string;
  name: string;
  filename: string;
}